requirements.txt
st_gsheets_connection==0.0.3
streamlit==1.28.2
streamlit_authenticator==0.2.3
matplotlib==3.8.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
streamlit==1.28.2
pytest==7.4.0
python-binance==1.0.19
pandas-ta==0.3.14b0
ta==0.11.0